Support long NAMES list
NAMES could span multiple messages for channels that have lots of users.
We don't need to clear our cached NAMES on each RPL_NAMREPLY; just do it
before the triggering WHOIS in sync() (which we already do).

Note that RPL_NAMREPLY is automatically sent on each successful JOIN to
a channel so we only have to do the sync()/WHOIS song and dance after
the first CHILD.

Another interesting note is that when an unsuccessful JOIN occurs in the
first CHILD: CHANNELS[channel] == 0 and there would be no
NAMES[channel, ...] entries. (there must always be at least one
NAMES[channel, NICK] entry for each successful join)

I don't think there's presently any code that relies on testing whether
the bot is in a certain channel, but in the future it should rely on
CHANNELS[channel] > 0.